add type check, path alias
