+++ data/ts/tidy/localize/index.d.ts
@@ -1,0 +1,12 @@
+type TLocalizeEtc = {
+  Key: number;
+  NameKr: string;
+  DescriptionKr: string;
+  NameJp: string;
+  DescriptionJp: string;
+};
+
+type TLocalizeEtcExcel = {
+  Key: number;
+  Bytes: TLocalizeEtc;
+};
