+++ data/ts/tidy/parcel/equipment.ts
@@ -1,7 +1,7 @@
 import type { EquipmentExcel } from "../type.js";
 import { Excel, cache } from "../../util.js";
-import { Localize } from "../localize/index.js";
-import type { TidyParcelItem } from "./index.js";
+import { Localize } from "../localize/main.js";
+import type { TidyParcelItem } from ".";
 
 export class Equipment {
   static table(): EquipmentExcel[] {
