+++ data/ts/tidy/parcel/furniture.ts
@@ -1,7 +1,7 @@
 import type { FurnitureExcel } from "../type.js";
 import { Excel, cache } from "../../util.js";
-import { Localize } from "../localize/index.js";
-import type { TidyParcelItem } from "./index.js";
+import { Localize } from "../localize/main.js";
+import type { TidyParcelItem } from ".";
 
 export class Furniture {
   static table(): FurnitureExcel[] {
