+++ data/ts/tidy/parcel/index.d.ts
@@ -1,0 +1,7 @@
+import { Rarity } from "../type.js";
+
+export type TidyParcelItem = {
+  Rarity: keyof typeof Rarity;
+  Localize: string;
+  Icon: string;
+};
