+++ data/ts/tidy/parcel/main.ts
@@ -1,15 +1,9 @@
-import { ParcelType, type Rarity } from "../type.js";
+import { ParcelType } from "../type.js";
 import { fatal } from "../../util.js";
 import { Currency } from "./currency.js";
 import { Equipment } from "./equipment.js";
 import { Furniture } from "./furniture.js";
 import { Item } from "./item.js";
-
-export type TidyParcelItem = {
-  Rarity: keyof typeof Rarity;
-  Localize: string;
-  Icon: string;
-};
 
 export class Parcel {
   static convert(t: keyof typeof ParcelType, id: number) {
